+++ apps/backend/src/libs/modules/server-application/server-application.ts
@@ -1,6 +1,7 @@
 import { config } from "~/libs/modules/config/config.js";
 import { database } from "~/libs/modules/database/database.js";
 import { logger } from "~/libs/modules/logger/logger.js";
+import { socketService } from "~/libs/modules/socket/socket.js";
 import { token } from "~/libs/modules/token/token.js";
 import { activityController } from "~/modules/activities/activities.js";
 import { authController } from "~/modules/auth/auth.js";
@@ -39,7 +40,7 @@ const serverApplication = new BaseServerApplication({
 	config,
 	database,
 	logger,
-	services: { userService },
+	services: { socketService, userService },
 	title: "TrackMates",
 	token,
 });