+++ apps/backend/src/modules/courses/courses.ts
@@ -2,7 +2,11 @@ import { logger } from "~/libs/modules/logger/logger.js";
 import { openAI } from "~/libs/modules/open-ai/open-ai.js";
 import { courseSectionRepository } from "~/modules/course-sections/course-sections.js";
 import { UserModel } from "~/modules/users/user.model.js";
-import { udemyService, vendorService } from "~/modules/vendors/vendors.js";
+import {
+	edxService,
+	udemyService,
+	vendorService,
+} from "~/modules/vendors/vendors.js";
 
 import { CourseController } from "./course.controller.js";
 import { CourseModel } from "./course.model.js";
@@ -10,6 +14,7 @@ import { CourseRepository } from "./course.repository.js";
 import { CourseService } from "./course.service.js";
 
 const vendorsApiMap = {
+	edx: edxService,
 	udemy: udemyService,
 };
 const courseRepository = new CourseRepository(CourseModel, UserModel);