+++ apps/backend/src/modules/vendors/libs/types/edx-image.type.ts
@@ -1,0 +1,5 @@
+type EdxImage = {
+	uri: string;
+};
+
+export { type EdxImage };